move pickup_serializer files
