+++ definable_serializer/contrib/pickup_serializer/viewsets.py
@@ -2,7 +2,7 @@ from rest_framework import viewsets
 from rest_framework.mixins import DestroyModelMixin
 
 from .views import PickupSerializerGenericView
-from definable_serializer.mixins import (
+from .mixins import (
     CreatePickupSerializerMixin,
     RetrievePickupSerializerMixin,
     UpdatePickupSerializerMixin,
