+++ definable_serializer/tests/for_test/urls.py
@@ -1,7 +1,9 @@
 from django.conf.urls import url, include
 from . import views as for_test_views
 
-from definable_serializer.routers import PickupSerializerRouter
+from definable_serializer.contrib.pickup_serializer.routers import(
+    PickupSerializerRouter
+)
 
 router = PickupSerializerRouter()
 router.register(r"answer", for_test_views.AnswerViewSet, "answer")
